requirements.txt
streamlit==1.20.0
pydub==0.25.1
PyAudio==0.2.13
playsound==1.3.0
plotly==5.13.1
requests==2.28.2
Pillow==9.4.0
jsons==1.6.3
streamlit-lottie==0.0.3